pyproject.toml
[project]
name = "enext_exchange_py"
version = "0.1.0"
description = "Wrapper for Euronext API"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "parsel>=1.11.0",
    "pycryptodome>=3.23.0",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
    "ty>=0.0.18",
]

[tool.pytest.ini_options]
minversion = "9.0"
addopts = [
    "-ra",                    # Show summary of all tests
    "-q",                     # Quiet mode
    "--strict-markers",       # Error on unknown markers
    "--strict-config",        # Error on config issues
    "-v",                     # Verbose output
]
testpaths = [
    "tests",
]
asyncio_mode = "auto"

[tool.ruff]
exclude = ["__init__.py"]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]
